Renamed Sdk > Proxy
Since the project contains proxy-specific stuff that is not
at all related to mocking (i.e. it has no concept of argument
matching, mock behaviors or configuration, recording of invocations
and the like), it's better to keep it highly focused and single
purpose, so that it can potentially be used for things other
than mocking, since it's a generic proxy library with minimal
API and footprint.