C# and VB support for proxy generator
Cleaned up the implementation significantly, moved as much as
possible to SyntaxGenerator-based codegen to be language agnostic,
but still generate C# as nice as possible.

Leverage MEF more and centralize retrieval of language services
in the new LanguageServiceRetriever. Our rewriters also leverage
MEF to register and filter themselves at generation time depending
on the project language.