Remove generator composition and use workspace service instead
Previously we would have our own external composition. That caused problems
since they may not be in sync with services in the workspace where the project
being processed lives. So we turn our language service retriever into a
workspace service instead and retrieve it from there.

Also removed all formatting since that's a responsibility of whoever is going
to consume the resulting document, and such formatting would only be needed
if the document is going to be written to disk or otherwise shown to the user.