Add basic proxy discovery
Discovers factory methods annotated with the ProxyGeneratorAttribute, which signals the
discoverer that generated proxies will be needed for the TypeArguments of the method
invocation.

The discoverers are per-language, so we leverage the project-specific language service
functionality to expose the discoverers. The driver ProxyDiscoverer invokes the per-lang
discoverer for each document and makes sure the resulting list is unique.

This adds a basic test for the functionality for both C# and VB.