Simplify and unify implementation of proxy discoverer
Instead of inheriting a walker for C# and VB, we can simplify
the code by using the LINQ API instead and since it's agnostic
to the type of nodes, we can centralize the implementation for
both languages by just parameterizing it to the concrete type
of syntax node for each language.

The rest of the discoverer is based on looking up the symbol
for the invoked methods, and that is generic too, so we don't
need to have two implementations that are mostly copied code.