Added basic MSBuild-based generator for use in projects
The Sample.CSharp.csproj project is used to verify things work, but
virtually all of the invoked code is already covered by unit tests.
The app imports the targets from the Console project, which has a
Debug-config-only project reference to the Generator.Build project,
so that the targets end up alongside the tool and the task assembly
for easier testing and debugging.

The approach chosen was to separate the proxy generation into a console
tool (`pgen`) that allows us to better isolate assembly loading issues
which plague Rolyn-based code generators (since there is already a version
of Roslyn in the process, from the compilers). This means performance
will not be great, but allows us to move forward instead of investing
more in workarounds. This should at some point be fixed by MSBuild itself
(see https://github.com/Microsoft/msbuild/issues/1754). See also
https://github.com/dotnet/roslyn/issues/17705 for other binding related
issues that won't exist at all if invoking an external tool.

We'll evaluate performance and come back to this if necessary.