Switch to using response files for -r and -s arguments to pgen.exe
When large numbers of references or source files are passed in to
pgen.exe (i.e. the sample project where we had EF + ASP.NET), the
command line buffer is too short and the arguments get truncated.

The standard way of doing this in MSBuild is to switch to response
files for the arguments that can cause this problem, namely -r,
-s and -a. A response file is basically an extra argument to the
pgen.exe of the form `@PATH`, where the `PATH` is a temporary file
written by the `ToolTask` base class whenever the `GenerateResponseFileCommands`
returns a non-empty string.

The file paths in the response file do not need the quotes like the
cmdline args, since they are read directly from the file. For
parsing, we use the same OptionSet used for the main args parsing.