Use AdditionalInterfaces to inject IMocked on all Moq proxies
There is an extensibility point in the proxy generation tool that
allows passing the full type names of symbols (that has to be resolvable
from the project where codegen is being performed) to inject as
part of the proxy generation.

This allows proxy behaviors to later on respond to those additional
interfaces to provide built-in behavior, such as Moq does for providing
mock introspection via the IMocked interface.