Add end to end sample and allow easy F5 on tasks and console apps
Of note: the sample projects contain a Mock.cs that contains the generic overloads
that pass along the current assembly to the actual implementation in Moq.dll. This
type will eventually be generated/included by the Moq targets file itself, since
it's not going to change per test project.

The reason we need it in the test project is that it makes it easier to find the
proxy types without costly discovery via attributes or reflection.