Improve code organization and improve document visitors approach
Previously, the ProxyGenerator.cs had knowledge of the order in which
to call the various codegen components, making it more complicated than
necessary.

This commit organizes all codegen components around the concept of an
IDocumentVisitor, which are registered in "layers", as supported by
Roslyn MEF. The `Scaffold` layer is the initial implementation that
leverages Roslyn code fixes to scaffold abstract class and interfaces
default implementations. The `Rewrite` layer runs after that, and
modifies the code so that it contains the actual proxy pipeline-based
implementation instead.

The folder structure has been updated to match this new logical
organization too.