Move rebuild to .cmd only
After a Clean, a Restore is required, so just doing if from the same
MSBuild script doesn't work, since projects need to be reloaded and
they won't in between restore and build.

Also added an /all switch that allows clean/restore/build/test to
all be run in a single call.