Fix broken tests since now the IMock implementation doesn't implement IMocked anymore
It doesn't make sense for the IMock to implement IMocked itself too, so we removed that
in a previous commit. But the tests were failing because the overload resolution on
AddMockBehavior was attempting to cast the mock as an IMocked, which previously did
work. So to avoid this, we provide direct overloads for AddMockBehavior and InsertMockBehavior
that extend IMock, just like for IMocked.