Shorten matchers names, fix Returns  bug
Remove 'Argument' from IArgumentMatcher implementations since
there aren't any other kind of matchers ATM. Also renamed
ArgumentMatcherFilter to SimpleBehaviorFilter to denote that
this is the simplest possible implementation but that others
can equally easily be implemented (i.e. for sequencing).

Returns had a bug in that it used a Stack<T> for matchers, but
these were Pop'ed  for ascending invocation argument index, causing
the exact reverse order of matchers to arguments mapping.