Refactored mock behavior filter and matching
Previously, each extension method had knowledge of how to push
matchers, how to build a filter from the contextual ones, and
it was getting tricky with potentially multiple extension methods
(i.e. multiple .Callback + .Returns calls) to follow when the
current invocation filter could be cleaned, and how to fetch it
from a subsequent extension method.

It was also very hard to debug and determine what filters were
in place, since they were all just Func delegates.

The new approach involves the Matchers, which provides Push and
AppliesTo method to deal with matchers and the resulting filter,
as well as a new strong-typed InvocationFilterBehavior which
holds and renders its inner InvocationFilter for proper debugging.

In addition, anonymous behaviors both for proxy as well as mocks
can now be given names that are used to override ToString for
both so they are more useful when diagnosing. We may want to
explicitly add something like an IName interface on top of both
interfaces, but it isn't needed for now.