Refactor Sdk to align with Proxy changes for conditional behaviors
The SDK can now focus on the argument matching and method setup, rather
than also implementing its own mini-pipeline inside a single MockBehavior.

Turned the concept of an InvocationFilter into a first-class IMockSetup
so we can later on provide facilities for retrieving configured behaviors
for a given setup, implement setup overriding and others.

The way MockInfo is initialized from the mocks changed now since the
list of behaviors is no longer IMockBehavior-specific, but rather the full
IProxy.Behaviors, so the codegen had to change too. This was an area that
was lacking unit tests of any kind (C# and VB) so we added two that ensure
that the codegen is correct.