Update to sub-pipeline per setup
The "outer" (proxy) pipeline is now extended with one particular
type of proxy behavior, the IMockBehavior which is a sub-pipeline of
additional behaviors that don't have their own AppliesTo. Instead,
behaviors can be manipulated just like for the entire proxy, but for
a particular "setup".

The setups are composed of: invocation target, invoked method and
invocation arguments. The setups support properly equality/hashing
comparisons so they can be uniquely identified within a mock.

In order to make this behavior on top of the existing proxy behavior
possible, the IProxy.Behaviors was changed into an observable collection,
and the MockInfo now tracks added/removed proxy behaviors to properly
index those that are IMockBehavior for later accessing by setup.

IEquatable<T> and IStructuralEquatable was implemented across all
matchers and method invocation implementation to support this new
indexing and uniqueness check for mock behavior setups.

Also made most mock introspection information "named" so they can
be tagged appropriately for better troubleshooting.