Adds support for generating proxies at design-time from the IDE
Adds analyzer and codefix to detect ProxyGenerator-annotated method invocations
and generates and adds a proxy for them in the IDE via a codefix.

This required a minor refactoring of the proxy generation we had, pretty much
everything is reused as-is.

Created a separate Proxy.Sdk that contains the document visiting/generation
interfaces so that third parties can extend this process via MEF in the IDE
natively, or via the existing MSBuild-based mechanism for the compile-time
generation and discovery.