Don't apply code fixes while visiting a doc
Ultimately applying the changes in the doc should be a responsibility of the
caller that's transforming the doc. In a codefix (IDE-based generation of the
proxy), this method is called to display a preview of the changes, so we most
certainly don't want to force it to apply to the current workspace (the opened
solution!) before the user actually commits the changes. VS handles that for
us automatically.

Also, since the IProxy interface is added at the Rewrite phase after the other
types have been visited, it's not necessary to skip applying code fixes for it
since there will never be any.