Place IProxy at the top, pass it to MockInfo/IMock
Rather than passing down the BehaviorPipeline's behaviors, pass down the
proxy itself to the MockInfo/IMock, so that if in the future we extend
IProxy, we don't have to modify generated code.

Change IProxy.Behaviors to be ObservableCollection<IProxyBehavior> instead
too, so that the mock (and third parties) can listen to behavior changes
on a proxy for their own purposes.