Renamed Moq.Proxy to Stunts, removed VSIX-based analyzer and refactored Roslynator inside
Roslynator was a nice idea but it was causing too much friction: we need to be able to
quickly iterate on the parts that consume Roslyn internals as easy as anything else. So
I copy-pasted what I needed into the new Stunts.Internal which is the one that has friends
visibility with Roslyn and has a very concrete and reduced API surface.

The anaylzer and code fix has been updated so that it can work solely via an <Analyzer
reference in MSBuild, which will make it far easier to update in the future (rather than
a VSIX). Turns out that we can't depend on anything MEF for codefixes if we want to use
them that way (thanks @sharwell for the link to the Roslyn source that clearly shows
this at https://goo.gl/DTGKrm).

Added stunts-specific samples in VB and C# for easy demoing.

Generating and updating stunts is 100% working now. Most likely, the build-time
generation will be relegated to a v2 release since it looks like the source code
generators may be ready by the time we need to tackle that and it will be far easier
than figuring out a well performing way of doing it with what's provided OOB right
now.