Flag as an error the presence of nested types
This is a stop-gap meassure until we can figure out why the ImplementInterfaces codefix
is not returning any code actions for the CSharpScaffold (maybe VB too?) when the type to
be implemented is a nested interface :(. In the IDE, the code action is properly available after
generating the (non-working, since the interface isn't implemented) mock class.
One possible workaround to not prevent the scenario altogether would be to leverage the
IImplementInterfaceService ourselves by duplicating the (simple) behavior of the CSharpImplementInterfaceCodeFixProvider
(see http://source.roslyn.io/#Microsoft.CodeAnalysis.CSharp.Features/ImplementInterface/CSharpImplementInterfaceCodeFixProvider.cs,fd395d6b5f6a7dd3)
and wrapping the code action in our own that would run the rewriters right after invoking the built-in one. I fear,
however, that the same reason the codefix isn't showing up when we ask for it, will cause that service to return
an empty list of code actions too.
It may be costly to investigate, and it doesn't seem like a core scenario anyway.
Another workaround might be to let the user manually implement stunts/mocks and detect the presence of
the IStunt interface and offer the rewriting codefix. This might be interesting too in general.

Related to #11.