Added recursive mocks codegen support
The analyzer can now specify if it wants recursive codegen to be applied.
Moq opts-in to this and the base classes do the work of automatically
also generating (and updating) the code for all generated classes for
interceptable members of the main type(s).

The performance impact is minimized by persisting the analyzer information
about what symbols need generating/updating in the reported diagnostics
themselves, so that the codefixes can simply execute on them.

This also simplifies the codefix code since it doesn't need to do any
more analysis on them.

Updated the sample so that performing calculations in the calculator
memory is done through a property that can be recursively mocked.