Switch to per-setup recursive mock codegen
Instead of whole-sale codegen of recursive mocks from the Mock.Of invocation,
only do granular recursive mock generation whenever a setup is detected that
accesses the mock recursively. This is far cheaper than generating for entire
object graphs, which can not only be very costly, but also not correct and
result in unnecessary or even non-compiling code to be generated.

Pending actual mock setup behavior to pick up the recursively generated types.