Centralize fluent mock context values in MockContext
Having a mixture of IMock.LastSetup, MockSetup.Current and the
CallContext<T> was confusing. Now all state is preserved using
CallContext<T>, but accessed uniformly via MockContext. This
allows us to add more fluent context values if necessary down
the road, without introducing more types to learn.

Reducing the IMock to just introspection (and not setup, as
the previous LastSetup did) also polishes the API and makes
it more single-purpose, which is good.