Add IMock<T> support and retrieving invocations for a lambda
This allows to easily retrieve a mock's invocations that match
a lambda, using the same mechanism we use for `Setup`. Had to
add `IMock<T>` so that the lambda can be typed. The implementation
is simply a wrapper, since we don't really need a typed MockInfo
that would otherwise complicate the codegen.