Add minimal Verify support
The combination of Verify with a simple lambda, plus an exception that
contains the `IMock` and the `IMockSetup` so that a user can further
validate the result.

The default Verify will check that there is at least one invocation
that matches the setup. Alternative overloads can be trivially implemented
by third-party extensions.