Add support for setting ref/out arguments via Returns
This involves tweaking the DefaultValueBehavior so that it does not
overwrite ref arguments (they are user-provided).

Also, a new Returns overload allows accessing the `IArgumentCollection`
for the current invocation, so that it can be updated.

Pending adding the same feature to the fluent interface of `Setup`
for void methods (although it would be a strange design choice to
have a void method with out args, but ref might be more common?).