Added support for typed ref/out returns
Other mocking libraries use an untyped dictionary of parameters to
set ref/out arguments. We support that model, but we felt that it
was not very type-safe and certainly not amenable to refactorings
and compile-type safe.

A mechanism by which you set up the returns by just referencing the
target method (as a method group) provides with the compile-time
safety. The problem is that the built-in Func/Action delegates
cannot be used for ref/out signatures, so typically you need custom
delegates to specify the callback in a typed fashion. This is quite
cumbersome and error-prone, and also very much not discoverable at
all.

So in addition to introducing type-safe ref/out via method groups,
we also provide a code fix that will automatically generate the
required delegate and set up the generic Setup call as well as a
boilerplate Returns with the right parameters.

This means all a user needs to provide is:

```
var mock = Mock.Of<IFoo>();

mock.Setup(mock.TryBar);
```

And a code fix will automatically be offered to turn that into:

```
var mock = Mock.Of<IFoo>();

mock.Setup<TryBar>(mock.TryBar)
    .Returns((ref string input, out Bar bar) => throw null);

...
delegate bool TryBar(ref string input, out Bar bar);
```

The user only has to replace the `throw null` expression with something
that initializes the `out` parameter and potentially change `ref` ones
as necessary.

Support for both VB and C# is provided.