Remove local temp package source to avoid restore errors
This never happens if the build.proj is run first, but it's
confusing for contributors to now be able to restore or
build the sln straight from a clone. Since adding that source
it typically a way to test the produced packages, it's better
left out to be added by whatever local projects you're using
for testing the proposed changes.

Fixes #21