Boost performance by reusing MSBuildWorkspace in project metadata reader
The external process should be longer lived so that the MSBuildWorkspace.Create
call (by far the most expensive since it needs to set up a whole new MEF host
with all Roslyn services, in addition to resolving the MSBuild framework assemblies
and all SDK-style targets and assemblies too) can be reused across calls to
ReadProject (renamed to OpenProject).

By switching to the JsonRpc style, the external process can live as long as the
build (or the AppDomain?) and maintain a single instance of the MSBuildWorkspace,
which then is reused across OpenProject calls. The RPC interface exposed allows
creating new workspaces on demand (disposing the existing one automatically, so
that it can live for the entire duration of the VS AppDomain eventually.

In addition, the console app now triggers a background resolving of all the
relevant assemblies and MEF eagerly, so that by the time the actual project
reading happens, all of the Roslyn components have already warmed up.

This combination of optimizations brought down the time for reading the test
project and all its project references from 20+ seconds (with the single-use
commandline for each project) to ~5 seconds (not counting the warm-up period
of ~2 seconds).