Prevent analyzer warnings during regular builds
At the moment, the analyzers are intended to be used only from the IDE.
In order to only add them to the compilation when the IDE requires them,
we only add the analyzers for design-time builds. This allows VS to
discover the analyzers, but does not add them for regular builds.

The nuget packages that provide these analyzers will also need to do
it conditionally via targets, instead of just exposing them in the
`analyzers` folder. Otherwise, the same warning/error would manifest
when installing the package.