Rename IMockBehavior > IMockBehaviorPipeline and IBehavior > IMockBehavior
This new naming is much more intuitive, since the "mock stunt behavior"
is a sub-pipeline that runs when the setup matches, so calling it a
mock behavior (to follow Stunt's behavior naming, since it implements
IStuntBehavior) *pipeline* denotes that it also has its own list of
mock-specific behaviors, which are now the IMockBehavior. The only
difference between IMockBehavior and IStuntBehavior is that the former
doesn't have an AppliesTo, since that responsibility has moved to the
IMockBehaviorPipeline which delegates its implementation to the
IMockSetup.AppliesTo.