Fix analyzer errors shown as warnings during build
The analyzers should be included via targets that do so only for
design-time builds (DTBs). To achieve this, we add TFM-specific
targets to the packages that add them.

Also fixed improper packaging of Moq API as netstandard1.3 when
it's actually 2.0, and changed Moq SDK to NS2 too to match Moq
itself (simplifies packaging).