Enable analyzers during build, codefixes at design-time
Running the analyzers for both Moq and Stunts is key to ensuring the
resulting code won't fail at run-time. As such, we want to ensure they
run as part of the regular build too, not just the IDE builds. This was
previously failing since the Workspace APIs that the code fixes use
could not be resolved during a normal build, but they resolve just fine
in the IDE build (design-time build, a.k.a. DTB) since the Roslyn
assemblies are binding-redirected by VS itself, but not MSBuild.

In order to fix this, we split the analyzers into a separte assembly,
`[Stunt|Moq].CodeAnalysis` and leave the code fixes in another one
`[Stunt|Moq].CodeFix`. Given this, only the code fixes need to dynamically
add themselves as analyzers for DTBs, also simplifying the packaging of
the analyzers.

Updated the packaging projects and samples to account for this too.