Fix improper unit test for ref/out recursive mocks
The old test had Debug.Asserts, instead of proper xunit asserts. This
masked an issue introduced when we added the MockTrackingBehavior automatically
via the DefaultMock, since there were duplicate tracking behaviors that
interfered with each other (since the current invocation setup had already been
determined).

After fixing that issue, I also noticed that the matching for parameters was
not being properly done for ref arguments (not comparing by the ref "element"
type), which would also cause issues. So the new TypeExtensions.cs was added
and used in the places where we did the check for byref + element type.