Align codegen with new targets that link files
We can now reliably locate the files always by their logical folder
name, since they are linked (even if hidden), so we can simplify the
probing code.

Also, we fine-tune the creation of the target folder in the project
to only the StuntCodeAction prior to adding the document to the project.