Improve default value support
Add ValueTuple support, fixes #12.

Move code to Moq.Sdk, since the very concept of returning default values from an invocation is closely tied to mocking, rather than a general concept used in proxies. It was earlier introduced in this layer because it made testing stunts easier, but instead of that, we keep the proper layering and just link the DefaultValueBehavior and the (renamed) DefaultValueProvider from the Stunts.Tests project.

The renaming from DefaultValue > DefaultValueProvider makes the class more discoverable, and preserves an existing API in v4.

Added support also for IQueryable, now that the default value classes are in a NS2 library.