Add mock configuration facade
Added a new `.AsMoq()` extension method that allows easily configurating of both the `MockBehavior` as well as the default value factories that apply to the mock.

Renamed the `GetMock()` from the Moq.Sdk to `AsMock()` also for consistency.

Note that the `AsMock()` exposes Moq.Sdk-only members, whereas the `AsMoq()` (and you could envision similar `AsFakeItEasy()` or `AsNSubstitute()`, say) provides a friendlier API that is mocking library-specific. We still expose the underlying Sdk introspection `IMock` via the `IMock.Sdk` property