Improve performance and equality comparisons
Switched to a HashSet<T> for storing invocations, which requires a fast(er) hashing that we borrowed and adapted from the BCL. This new HashCode class allows us to not depend on `Tuple` for doing the structural comparisons so we also did away with the `IStructuralComparable` interfaces in favor of the "regular" `IEquatable<T>`.

Added missing `MethodInvocation` equality comparison test as well as constructor argument validation that we do now.