Drop usage of `InternalsVisibleTo` hack into roslyn internals
We were using the approach of signing our Stunts.Internal project with
the same key used in a roslyn test assembly, which gave us internals
access with full intellisense and compile-time checking. We only needed
few internal members access, though, and this was a risky move anyway.

To prove the risk, v2.11 (VS2019) actually moved one of the internal
types we used to another namespace, and moreover, changed entirely the
key used to sign the test assemblies, so the entire approach doesn't
work anymore.

This commit reverts that approach with one based on limited reflection,
which might be slower but it's only used during codegen. In order to
guarantee that we don't break on future versions of roslyn, a unit
test that fetches the latest stable and pre-release versions of the
Workspaces.Common assembly (the one that contains the internals we
need) from nuget.org, builds a project on the fly with that as a
dependency, and tries to resolve the internal members. If any fail,
a test will fail, so we'll know ahead of time we need a workaround.
It has already proved its usefulness by uncovering the move of one
of the methods we used to another static helper class :).