Fix dynamic mock generation and enable test
There was a compilation issue caused by the use of `RefExpression`
which resulted in a compilation error saying `ref` was an invalid token.

By replacing that expression using the other overload, we could fix the
compilation error, although now the generated code will be slightly
different but have no semantic change (replaced a regular argument with
a named argument now).