Initial simple migration from new Mock<T> to Mock.Of<T>
This simple migration (which would obviously not work in general since
it's not replacing the variable declarations, nor migrating the member
invocations on Mock<T> that would now be extension methods or arranging
of overloads and the like), serves to set up the way we'll author and
test these migrations.

The migration APIs will live in the Moq/Migration folder, so we know
they are there specifically to light up the analyzer/code fixes. All
types and members will be marked `EditorBrowsable(Never)` and throw
`NotSupportedException`. The code fixers should support batch fixing
to a few passes of code-fixing should leave the user with a fully
working v5 code base. This approach makes the migration points and the
new APIs easier to pinpoint and learn, unlike a command line or global
migration tool.

We may still provide a command line tool to migrate an entire solution,
maybe as a dotnet global tool?