Make sure we add labels for branch and pr
We don't want to publish CI packages that can not be distinguished from
each other, since build metadata cannot be used to disambiguate.
Therefore, each PR should have its number as part of the label on the
commit.

The same issue happens across feature/dev branches, so we also append
the branch name to the label.