Simplify local dev nuget consumption for easier testing
Update the `PackageOutputPath` so it points to a single folder
(`out` at the repo root) so that testing locally is simply a
matter of setting up that path as a source.

Also simplified the package cache cleaning target since deleting
and copying to %TEMP% isn't very useful since that folder is cleaned
across restarts and VS fails to restore afterwards if the entire
`packages` folder is missing from there by the time a solution is
opened.

Document this in a section in the Readme too.