Change StuntBehavior/MockBehavior factory types into regular classes
MockBehavior in particular was a duplicate type in Moq.Sdk and Moq (an enum here for Loose/Strict) and it was confusing. Since we had to eliminate this duplication in Moq by creating the new DelegateMockBehavior with regular constructors (replacing the previous MockBehavior.Create overloads), we also aligned the StuntBehavior so the API is symetric. As shown in the updated tests, there wasn't much loss in conciseness anyway.