Make IMoq inherit IMock, which is more intuitive for instrospection/configuration
Instead of exposing a rather cryptic `Sdk` property on `IMoq` to access the introspection
API, make it just inherit from it, since they are likely to be used together for advanced
scenarios anyway.

A new `MockDecorator` class makes it easier to wrap the built-in `IMock` from the actual
target object into a typed interface that aggregates the base with library-specific
configurations.