Fix analyzer errors raised in csc for consuming projects
For some reason, a mismatched version of the VB code analysis assembly causes the
analyzers to not be loaded at all. It does not seem to be a problem for the C#
assemblies, which is quite puzzling.

For now, include the `Microsoft.CodeAnalysis.VisualBasic` assembly so that codegen
works. At the `AutoCodeFix` level things work because it performs its own assembly
resolution, though.

Also remove unused old roslyn internal dependency assembly.