Make VerifyException render more nicely
This exception was not passing any message to the base class, therefore a default message was being rendered which wasn't too intuitive. So we add the typical overloads receiving message and optional inner exception, and pass the former as the string rendering of the setup (that failed verification) as an initial simple implementation that is already better than before.