Simplify targets and make them self-contained, implement clean
The existing `Clean` support was cumbersome and based on some complicated
details of the actual implementation of the clean targets as explained in
https://github.com/Microsoft/msbuild/issues/2408#issuecomment-321082997.

However, our needs are much simpler, since the intermediate output path for
both stunts and mocks is well-known and generally not user-specifiable. The
(new?) `@(Clean)` item group should be enough for our needs in that regard.

Also, Moq targets were depending too tightly on details of Stunts targets,
which added complexity for little gain. So instead we simplify both by using
the same approach to adding files to `CoreCompile`, `Clean` and `@(CodeFix)`
analyzer inclusion.