Add new behavior for setup scope runs
When the pipeline is run in a `SetupScope`, we should not run the entire pipeline, but rather just the behaviors we know are safe for execution during setup, specifically `MockContextBehavior` (to set up the `MockContext.CurrentInvocation` and `MockContext.CurrentSetup`) and `DefaultValueBehavior` to return a default value for further fluent API invocations to work on.

This simplifies the behaviors since they don't need to check and bail during active setup scopes, and makes them more self-contained.