Simplify Behavior implementation by using SkipBehaviors
Instead of duplicating behavior and having the Behavior property modify the actual pipeline, leverage the new mechanism we added to allow skipping behaviors dynamically for the current invocation.

With this new feature at the pipeline level, we can create a new configuration behavior that can check the `MockBehavior` and simply skip the `StrictMockBehavior` if it's not `MockBehavior.Strict`. This greatly simplifies the implementation and makes both default and strict behaviors always present on the pipeline, but dynamically skipped based on configuration.

This requires introducing a mock state-level "field", which we now save with the key `IMoq.Behavior`, which should be the guideline for extensions that need to persist state in the mock: use the `[Type].[Property]` syntax to minimize chances of collision.