Ensure both analyzers and codefixers have NuGetPackageId metadata
The NuGet asset resolving target already adds the metadata to the `Analyzer` items inferred from the `analyzer` folder in the package, but the `CodeFix` we provided did not contain this metadata. This makes it harder for a consumer to actually filter/remove them if they want to customize codegen in advanced way.

This would also allow a project to replace the package-provided analyzers and code fixers with locally built ones directly from the bin directories of the Moq solution.