Add support for mocking generic types
We previously used the very rudimentary `MetadataName` property to determine the type
to generate a stunt for. This was error prone since it does not contain the full metadata
for generic types (or nullable types) and basically you cannot resolve an `INamedTypeSymbol`
from it.

Since we need to support a more reliable and robust way to communicate back from an
analyzer to its code fix what is the target type(s) to generate stunts for, we implement our
own formatting (leveraging the `SymbolDisplayFormat` and then implementing a resolver
that parses the C#-like generic expression (i.e. `IEnumerable<KeyValuePair<string, int?>>`)
and resolves it back to a fully constructed generic `INamedTypeSymbol`.

We leverage the Superpower library for the parsing which is fast, provides great error reporting
and is a breeze to use and understand.