Move OverrideAllMembersCodeFix to CodeFix assembly to avoid csc error
This code fix should only load in our own workspace, since csc does not contain the
required dependencies for it to load properly. This was causing transient build errors
when csc.exe would try to inspect the types via MEF and fail in the static initialization
of the RoslynInternals class which even if not used since the code fix was not exported,
was nevertheless initialized statically when reflected by the MEF composition.

By moving it to the CodeFix assembly, we also make it work transparently like all other
code fixers, which are referenced just by name. We were already doing that from the
C# and VB scaffold document processors, so no changes were necessary there.