Unify naming conventions for runtime lookup
Code generation support for generics introduced the `Of..` name to
append the generic types being mocked (i.e. `IEnumerableOfStringMock`).
This was not reflected in the run-time naming classes so some tests
were failing.

In addition, there was quite some duplication between MockNaming and
StuntNaming, which is now unified with the former delegating to the
latter all behavior and just providing different DefaultNamespace and
DefaultSuffix values.

The new implementation is more aligned with the NamingConvention used
for codegen now.