Drastically simplify As<T> support by adding new Mock<T...Tn>
Instead of the quite complicated implementation that tried replacing the
pipeline on a mock, instantiating a new one, cloning its state, etc.,
we will just provide an analyzer+codefix that will expose the `T` as
part of the generic `new Mock<T...>` constructor, which means the code
generator doesn't need to get any smarter, and we can actually help
migrate people to the newer pattern where all types are known to the
factory from the beginning.

In order to do this, we also brought back the constraint on the `T`
for `IMock<T>`, which aligns with v4 and also allows us to cast from
within the `As<T>` implementation. This constraint was added to all
APIs that exposed `IMock<T>`.